reorganization: move code into separate files
