+++ editor/src/constants.ts
@@ -1,0 +1,7 @@
+export const CANVAS_WIDTH = 1366;
+export const CANVAS_HEIGHT = 800;
+export const CANVAS_BACKGROUND = "black";
+export const PARTICLE_COLOR = "#FCE93D";
+export const PARTICLE_COLOR_FOCUS = "#D9CB40";
+
+
