+++ editor/src/utils.ts
@@ -1,0 +1,9 @@
+import { redrawCanvas } from "./canvas.js";
+import { refreshButtons } from "./dom.js";
+import { type State } from "./types.js";
+
+export function redrawAll(state: State) {
+  redrawCanvas(state);
+  refreshButtons(state);
+}
+
